requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
python-multipart